requirements.txt
flask
flask_restful
face_recognition_models
Click>=6.0
dlib>=19.3.0
numpy
Pillow
face_recognition
